0.7.7b Release
NOTE: There was a MAJOR speed problem with this version. DO NOT USE IT! This version sucked.

Changes:

Blocking specific addresses now possible. Edit the PHP file, and at the top is an array with the addresses you want blocked as elements. Currently you can block addresses in one of three ways:

  1.2.3.4     - plain IP address
  1.2.3.4/24  - subnet blocking
  php.net     - domain blocking

 The default blocked addresses are:

  10.0.0.0/24
  172.0.0.0/24
  192.168.0/16
  127.0.0.0/24

 These are blocked by default since they are reserved for local and internal purposes. If I have missed some, or if I have one of these mistaken, please email me.

 I haven't decided if I want to add IP range blocking yet. I personally can't think of any reasons why you would need it, but if you do, just email me saying that you would use it and I'll be motivated to add it. If you don't it probably won't get done, but there still is a slight chance that it might. I've got nothing better to do, but I'm just not motivated enough to do it at this point :).

Error displaying engine created, for blocked pages, connection problems, etc

Sessions are now used to store the encoding information for URLs and cookies. Before, the encoding string was stored in the cookies, which meant that it was possible for an administrator to see the encoded string as well as the decoding portion, then find the code on my site and decode it. This obviously is not a good thing, so I moved the key to decode the string into the session, and increased the size of that key from 12 to 30 characters long. Note, this does not affect the cookies of the sites you visit or the proxy's cookies at all, just makes sure that the encoding is secret.

Sessions time out after 24 minutes, so an administrator must see an encoded URL and send your cookies out to the site within 24 minutes of you visiting. You can change that time in the PHP config by changing the "session.gc_maxlifetime" setting. If you want me to change it by default, email me and I'll consider it.

New "Encode Cookies" feature, encodes the values of the cookies for sites using the same key as the "Encode URLs" uses, and is now in a session as noted above.

Changed the cookie separator to the cookie prefix, instead of "__proxifier_sep__".

Minor tweaks in parsing engine

Planned:

Code cleanup: the code is getting a little messy, I'm going to try to fix that. I might try OOP for some of it, I might not. I'm not sure yet. I'm definitely going to increase the amount of regular expressions, replacing the "substr"s and stuff.

URL form on the page as an option, and javascript enabled with a toggle to see it or not, degrading gracefully.