0.7.8.1b Release
Changes:

Fixed "Encode URLs" with Internet Explorer bug (IE doesn't have btoa().. so I had to write my own base64 encoding function .. which required me to write a binary to decimal conversion since JS doesn't have that.. and an exponential funcion because JS doesn't EVEN HAVE THAT! How JS has survived this long without an exponent operator is beyond me..)

Fixed up the look of the front page of the proxy: I hadn't tested Proxifier on IE for awhile (causing the above bug) and I didn't even realize that it looked so terrible in IE. I fixed most of the look, but there are still a few other problems. I also adjusted the look for every browser just a tad as well. I may eventually just make the front Proxifier page with CSS float and divs instead of a table... might fix a lot of the alignment issues.