0.7.9b Release
License of proxy changed from GPL to a Modified BSD. See http://bcable.net/license.php for more details on the new license.

Changes:

Default User-Agent is no longer to just not send it, but the actual User-Agent string sent to the server. It is still possible to change this field, and to remove it entirely, but this way many sites that require it are working by default. Detection of the user's platform and browser name is also done to inform the user what the proxy will report. Detection is minimal, however, and does not detect very many platforms as of now.

To reduce the amount of space taken by the Javascript of my script in the source, not only are the Javascript functions but also the Javascript regular expressions are found in separate requests to the server. This greatly cleans up the source code of a retrieved page, so that the user isn't completely confused when viewing the source.

Lots of speed improvements with this release.

Everything is now sent gzipped to the user, reducing the size of every file and speeding the communication to the proxy up dramatically.

All headers from the server, minus a select few, are sent to the user now, but obviously still not vice versa as that would be a privacy hazard. This way, certain caching mechanisms can work correctly reducing the amount of data flowing through the proxy.

Amount of regular expressions used were greatly reduced, including the minimum of 16 regexps per URL, and another 7 for each modification to the URL performed. Now, a flat 9 regexps total, regardless of modification, is being performed. This was a major CPU burner in the previous versions, and the overhead has now been reduced almost 10 fold for this section of code.

Persistent connections to the servers has been removed (IE: pfsockopen replaced with fsockopen again). It seems that the persistent connections with the servers were not helping at all in the slightest for speed, and in fact, were hampering it to some extent. Until I find a way to have the persistent connections time out after inactivity, this will remain. The connections weren't EVER expiring, which created many open connections that didn't need to be open.

The gap between the "fsockopen" and "fclose" has been reduced slightly so that the connection wasn't just sitting there waiting to be operated on, while the proxy script was doing some work on its side. This slightly reduced the amount of time the connection was left open.

Other engine fixes, speed tweaks, and code optimizations were also done.

Planned:

Possibly change "proxy.php?blah_url=<URL>" to: "proxy.php?<URL>". This would make the URLs look a bit nicer in the source, as well as make it easier to change URL by editing the address bar. Auto detection of encoded URLs, as well.

Have an option to insert a URL form at the top of every page. Possibly using javascript to fold up, as well as degrading gracefully for those who don't have/allow javascript.

Modification of the page's title, as well as masking of the "favicon", so that when minimized a page wouldn't look suspicious to shoulder surfers.

If not the above, a feature to "hide" the page you are looking at by an invisible button on the page which flips back and forth between a hidden page and the mask page. This may not be possible as the address bar will alway report the correct URL of the proxy, and using IFrames, you cannot access the "mask" page's title (access to an IFrame's document object is forbidden for obvious security reasons). Good idea, but hard if not impossible to implement. I'll try to think something up for it.