0.8.2b Release
Changes:

Fixed a bug regarding blocked addresses. Blocked IP addresses are compared on a binary level, but the binary was being assembled incorrectly. For instance, the binary number "00000101" was being represented as "10100000". Therefore, not only would "00000101" be blocked, but also something that actually was "10100000" would be blocked as well.

Added an option for changing the protocol the script is installed to. Defaults to "http", which would be what most people would need, but in the instance someone installs it on an SSL certified URL, it can be changed to "https".