1.0-rc1 Release
Changes:

PRIVACY FIX/INFORMATION LEAK: The Javascript "document.cookie" feature was not hooked before, causing an information leak to the site telling them ALL of your cookies defined through the proxy. This could cause some major security and privacy problems. This also fixed some functionality problems on certain sites that required the "document.cookie" feature to operate correctly.

PRIVACY FIX/INFORMATION LEAK: The Javascript "navigator" object was not hooked before, causing an information leak to a site telling them your actual useragent if you forged it through the proxy options.

Changed the format of the proxied URLs. No matter what, the URLs are always the same now. So you can FINALLY send links to your friends without there being a problem!

 Each URL now has a "mode". Please note, you do not need to know what the rest of this means, this is just a technical description for those who are interested. Instead of GET variables such as "force_main=1", "<COOK_PREF>_page_framed=1", "<COOK_PREF>_frame=1", etc, the following table describes what each URL points to:

  ?=

 This is the "null" mode. This is used if you have the persistent URL feature disabled, or for a page that isn't supposed to be framed (external scripts/styles, etc).

  ?=&=

 This is the "frame" mode. This will display the URL with the persistent URL frame.

  ?=_&=

 This is the "framed" mode. This is the framed page request, and is almost identical to the "null" mode. The only difference here is that it is labelled and used for targetting the "_top" HTML target in HTML and Javascript to keep the page framed correctly.

  ?=.&=

 This is the "child frame" mode. Any URLs loaded by a "framed" mode URL have this mode. This makes sure any URLs clicked in <iframe>s or <frame>s on a page stay within that frame, and any target attribute is dealt with correctly.

  ?=-&=

 This is the "menu" mode. This will force the page to display the main menu, and have the specified URL be inserted into the URL form.

The default title when loading a page used to be incredibly long and illegible. It was generally part of the URL of the proxy, and you had no idea what was loading. Now, the title is the actual URL entered in, and is later altered to be the correct title of the page.

Added support for "shortcut icon"s (also known as "favicon"s or "favicon.ico"), as well as fixed some other various bugs regarding how the title was changed.

New/updated useragents. Mac OS X added as an option (PPC) for Firefox, Safari, and Opera. Also added detection of BSD for the "Actual" option. The other useragents were updated to be the most recent versions of all the supported browsers.

Along with the new URL format, URLs are now cleaner looking in the address bar because they are no longer encoded.

Detection of where and when to decode a passed URL, in case a GET request does indeed encode a URL. The script looks at the first 6 characters, and if it has a '%' in that substring, then the URL is decoded. The first 6 characters SHOULD have a '%' in it if it is encoded no matter what.

Fixed a bug where the URLs and variables being passed via GET and POST would incorrectly parse ':', ' ', and other various special characters.

Removed code which added the GET variables to the current URL, which is unnecessary due to the new URL format.

The script now automatically sets the "error_reporting" PHP setting to 'E_ALL ^ E_NOTICE', making notices from PHP not displayed even on servers where the PHP.ini tells them to be displayed.

Added automatic detection to determine whether you are running on HTTP or HTTPS, and the URL is automatically adjusted accordingly. A manual override is available if necessary in the options at the top.

Fixed a bug where the script would parse URLs with quotes around them incorrectly.

Fixed a bug where the script would parse URLs even though they were supposed to be "locked". For instance, "javascript:" URLs were getting parsed.

Added support for flash movies that use the format: <param name="movie" value="MOVIE_FILE">

Fixed a bug where CSS "url()"s had &quot;, &#34;, or &#39; to signify quotes. Those HTML entities are apparently parsed before they are handed over to CSS, and the script now parses accordingly.

Adjusted regular expressions that parsed Javascript to ensure the command was starting at the beginning of a line or statement. This fixed many bugs involving various odd coding styles.

Fixed a bug where the end HTML comment before a "</script>" tag was incorrectly parsed, putting the "purge" statement in the incorrect place.

Fixed a bug where <meta> tags using "content" HTML attributes without quotes wouldn't work, and in the process, reduced the amount of regular expressions parsing <meta> tags by 2.

Fixed a bug where Javascript variables were not parsed if they contained dollar signs ($) in them, which apparently is legal in Javascript. (yuck!)

Fixed a bug with the "target" attribute of <form>s being incorrect, and the <base> tag not affecting the "action" attribute of a <form>.

Fixed a bug with the hooking of certain attributes of objects to have better logic in determining whether or not to redirect. This fixed many sites which seemingly redirected to a random page on their site for no reason, or got caught in an infinite loop.

Fixed a different bug with the hooking of the location attribute and it not getting set correctly.

Added proper Javascript hooking of the setTimeout() and setInterval() top-level functions.

Instead of the "innerHTML" attribute in Javascript being parsed by regular expressions, it is now sent through the Javascript hooking engine.

Adjusted parsing of the Javascript eval() top-level function to not be a direct hook into the function, but instead now uses regular expressionsto insure the eval is performed as a top-level function, fixing many many sites.

Added proper hooking and parsing for the Javascript "location.search" property.

Properties of various Javascript objects weren't being changed when called by sites. This means they were getting the proxy's URL, not the URL they were expecting. To fix this, a new function reverses the proxied URLs to their original values before the site's Javascript accesses them.

When Javascript was getting URLs, they were still encrypted if that option was enabled. So, to fix this, I also added an implementation of the base64 decoding algorithm (it already had the base64 encoding algorithm), and also translated the PHP function that decodes the encrypted URL into Javascript.

Javascript XMLHTTPRequest objects were originally hooked on the fly. This required the use of multiple lines of code to do so, and sometimes the XMLHTTPRequest was called in a spot where it required it to be on one line. Now, all of the hooking is done with one function call to fix this.

Many many many frame breaking bugs were fixed in Javascript.

The Javascript code in general was cleaned up a bunch, as well as new methods of hooking employed to reduce code and to make the code more flexible.

Instead of using PHP to echo all of the values of various script definitions directly into the Javascript code, it now converts all these values into Javascript at once, and then the Javascript code requests these variables instead of expecting values to be placed in the code directly.

Less, and more efficient, regular expressions are used to parse Javascript now.

The "REGEXPS" section of the code was practically unreadable in the past. I have added some comments to help other people and even myself in the future in order to understand what they all mean and do.

General Javascript and PHP optimization fixes.

Many many more bugs were also fixed, but could not be listed here because I forgot to write them down.