Testing suite updated.
Added an assert for the encryption options of the proxy, and separated the
'remove' assert into its separate components.  Also made overlapping
asserts to make sure some options play nice together.

Fixed a few more bugs in the sandboxing engine.

Fixed the proxy to allow for the sandboxing of the $_SERVER variable.

Fixed a bug with the proxy where it would attempt to decrypt any cookies
not related to the encrypted cookies anyway.  This probably slowed some
things down a bit.

Moved the calling of asserts into an array stored in an easier accessible
location (asserts/asserts.js).

Changed the formatting of the assert display slightly.