Testing suite updated.
Consolidated the markup code which is repeated iver multiple files into an
include directory which can be modified from one central location.

Added a new assert to test the standard operation of cookies.

Fixed the 'remove_cookies' assert to actually test if cookies were being
removed or not.

Fixed a small code formatting issue with the proxy code.