Regular Expressions and Asserts
Renamed all of the regular expression variables to be more logical in their
organization, and added documentation for every single variable.  This
greatly reduces the confusion surrounding this part of the script.

Fixed the setAttribute parsing code to work with the recursive regexps, and
re-enabled them.

Fixed a bug where Javascript was being parsed inside of comments.

Fused all the markup assert includes into one file, and made the Javascript
much more simple for the markup asserts.  Started work on more robust
Javascript asserts.

TODO

Proxy:

Figure out a way to not parse Javascript inside of multiline comments.
Perhaps the same way HTML is parsed: divide into multiple segments via
comment regexp, then parse small chunks at a time.  This could also be
faster (...or slower, should test that).

Take a look at $js_lochost (it's unused), and make sure it's not needed
before throwing it out.  Make sure location.host(|name) actually is parsed
correctly.

See if some of the object['attribute'] regexps overlap with the
document.attribute regexps, and see if there is a way to consolidate these.

Figure out how to solve the problem with SSL being enabled on paypal.com,
but when you visit paypal.com it is broken because it uses domains such as
paypalobjects.com as well.

Figure out a way to add a popup after submitting a form on an SSLd page
that has an action of a non-SSLd page.

Go to each section of the script, line by line, and make sure it is well
documented and each comment is standardized.

Testing Suite:

On the same line as the success/failure, state how long the assert took to
run.  Could use Javascript or PHP, depending on which is easier and better.

Look through the javascript_php assert line by line, make sure everything
is being parsed correctly.  Organize the Javascript a bit better and
provide comments for each line or section, and/or use separate asserts for
each thing.  Perhaps the same for markup as well.  See if markup.php can
get tossed (or use markup.html or something).

Need to make asserts (somehow) for Javascript parsing of Javascript.  Maybe
some mechanism that waits until it receives X signals from Javascript
before saying true/false.